Migrate OperationsStats
