+++ packages/services/api/src/modules/operations/resolvers.ts
@@ -1,162 +1,9 @@
-import { hash, nsToMs, parseDateRangeInput } from '../../shared/helpers';
+import { nsToMs, parseDateRangeInput } from '../../shared/helpers';
 import { createConnection } from '../../shared/schema';
 import { OperationsModule } from './__generated__/types';
 import { OperationsManager } from './providers/operations-manager';
 
 export const resolvers: OperationsModule.Resolvers = {
-  OperationsStats: {
-    async operations(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      args,
-      { injector },
-    ) {
-      const operationsManager = injector.get(OperationsManager);
-      const [operations, durations] = await Promise.all([
-        operationsManager.readOperationsStats({
-          organization,
-          project,
-          target,
-          period,
-          operations: operationsFilter,
-          clients,
-        }),
-        operationsManager.readDetailedDurationPercentiles({
-          organization,
-          project,
-          target,
-          period,
-          operations: operationsFilter,
-          clients,
-        }),
-      ]);
-
-      return operations
-        .map(op => {
-          return {
-            id: hash(`${op.operationName}__${op.operationHash}`),
-            kind: op.kind,
-            name: op.operationName,
-            count: op.count,
-            countOk: op.countOk,
-            percentage: op.percentage,
-            duration: durations.get(op.operationHash)!,
-            operationHash: op.operationHash,
-          };
-        })
-        .sort((a, b) => b.count - a.count);
-    },
-    totalRequests({ organization, project, target, period, operations, clients }, _, { injector }) {
-      return injector.get(OperationsManager).countRequestsAndFailures({
-        organization,
-        project,
-        target,
-        period,
-        operations,
-        clients,
-      });
-    },
-    totalFailures(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      _,
-      { injector },
-    ) {
-      return injector.get(OperationsManager).countFailures({
-        organization,
-        project,
-        target,
-        period,
-        operations: operationsFilter,
-        clients,
-      });
-    },
-    totalOperations(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      _,
-      { injector },
-    ) {
-      return injector.get(OperationsManager).countUniqueOperations({
-        organization,
-        project,
-        target,
-        period,
-        operations: operationsFilter,
-        clients,
-      });
-    },
-    requestsOverTime(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      { resolution },
-      { injector },
-    ) {
-      return injector.get(OperationsManager).readRequestsOverTime({
-        target,
-        project,
-        organization,
-        period,
-        resolution,
-        operations: operationsFilter,
-        clients,
-      });
-    },
-    failuresOverTime(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      { resolution },
-      { injector },
-    ) {
-      return injector.get(OperationsManager).readFailuresOverTime({
-        target,
-        project,
-        organization,
-        period,
-        resolution,
-        operations: operationsFilter,
-        clients,
-      });
-    },
-    durationOverTime(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      { resolution },
-      { injector },
-    ) {
-      return injector.get(OperationsManager).readDurationOverTime({
-        target,
-        project,
-        organization,
-        period,
-        resolution,
-        operations: operationsFilter,
-        clients,
-      });
-    },
-    clients(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      _,
-      { injector },
-    ) {
-      return injector.get(OperationsManager).readUniqueClients({
-        target,
-        project,
-        organization,
-        period,
-        operations: operationsFilter,
-        clients,
-      });
-    },
-    duration(
-      { organization, project, target, period, operations: operationsFilter, clients },
-      _,
-      { injector },
-    ) {
-      return injector.get(OperationsManager).readGeneralDurationPercentiles({
-        organization,
-        project,
-        target,
-        period,
-        operations: operationsFilter,
-        clients,
-      });
-    },
-  },
   DurationValues: {
     p75(value) {
       return transformPercentile(value.p75);
